Test dives for duplicate names and conflicting GPS
The conflicting GPS is somewhat artificial. Since the GPS gets parsed
first in our syntax, the different GPS location would create a new dive
site and we'd end up with two dive sites of the same name with different
coordinates. By having a location tag with just the name and no
coordinates we make sure that this gets identified with the existing dive
site and THEN add the GPS coordinates in the second location tag.

This would never happen in a XML file created by Subsurface, but it does a
good job in testing the different code paths.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>