Use rpath to make binaries run without LD_LIBRARY_PATH
When building binaries with our custom libraries it is always a pain to
ensure that all the libraries are found. Adding an rpath can help.

When running qmake with CONFIG+=setRpath we include the paths for our
private builds of libgit2 and libssrfmarble (assuming those are not using
the system libraries).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>