build.sh: successfully build an App bundle on the Mac
Some install name magic on two of the libraries we build was necessary for
things to work out correctly.

And I added an install step to the default build that puts Subsurface in the
install_root on other OSs and creates the Subsurface.app under
subsurface/build/Subsurface.app on the Mac.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>