Build script: use sufficiently new libgit2 and enable cloud storage
Turns out that as of a day ago or so tip of libgit2 master appears broken
(the in memory ssh key test in the cmake file fails). But the specific
commit that I'm picking here appears to work and is also new enough that
https and ssh based cloud storage works.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>